requirements.txt
django
python-dotenv
django-cors-headers
langchain
langchain-google-genai
langgraph
google-genai
ccxt
uvicorn
pandas
binance-sdk-derivatives-trading-usds-futures
